ema-futures16/ema-futures/requirements.txt
python-binance
numpy
colorama
requests
binance-futures-connector
